# fix(tasks): allocation tracker prevent counter overflow