@camchenry 2025-08-14
@overlookmotel I think it's a fair fix. Performance for this task isn't a high priority and it's fast enough for now.